requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
PyMuPDF==1.23.7
python-multipart==0.0.6
openai==1.2.3
python-docx==0.8.11
pytesseract==0.3.10
Pillow==10.0.1
pypdf==3.17.1
python-dotenv==1.0.0
docx2txt==0.8
numpy==1.26.1
firebase-admin==6.2.0
weasyprint==60.1
stripe==7.0.0
docling
pdf2image==1.16.3